+++ intermine-ios/Constants.swift
@@ -37,6 +37,8 @@ struct Icons {
     static let bookmark = UIImage.init(icon: .FABookmark, size: CGSize(width: 35, height: 35))
     static let login = UIImage.init(icon: .FASignIn, size: CGSize(width: 35, height: 35))
     static let close = UIImage.init(icon: .FATimes, size: CGSize(width: 40, height: 40))
+    static let menu = UIImage.init(icon: .FABars, size: CGSize(width: 40, height: 40), orientation: UIImageOrientation.up, textColor: Colors.white, backgroundColor: UIColor.clear)
+    
 }
 
 struct Colors {
