--- intermine-ios/MinesTableView.xib
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="12120" systemVersion="16E195" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES">
-    <device id="retina4_0" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="12088"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" id="iN0-l3-epB" customClass="MinesTableView" customModule="intermine_ios" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
-            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-            <subviews>
-                <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" scrollEnabled="NO" showsHorizontalScrollIndicator="NO" showsVerticalScrollIndicator="NO" style="plain" separatorStyle="none" allowsSelection="NO" rowHeight="60" sectionHeaderHeight="1" sectionFooterHeight="1" translatesAutoresizingMaskIntoConstraints="NO" id="eZd-xX-eav">
-                    <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
-                    <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
-                    <constraints>
-                        <constraint firstAttribute="height" constant="667" id="Ndx-EF-i13"/>
-                    </constraints>
-                    <connections>
-                        <outlet property="dataSource" destination="-1" id="Lrb-jk-BqK"/>
-                        <outlet property="delegate" destination="-1" id="PKr-wk-Cw2"/>
-                    </connections>
-                </tableView>
-            </subviews>
-            <constraints>
-                <constraint firstItem="eZd-xX-eav" firstAttribute="centerY" secondItem="iN0-l3-epB" secondAttribute="centerY" id="dHH-SM-1al"/>
-                <constraint firstItem="eZd-xX-eav" firstAttribute="width" secondItem="iN0-l3-epB" secondAttribute="width" id="gtZ-7A-Oqy"/>
-                <constraint firstItem="eZd-xX-eav" firstAttribute="centerX" secondItem="iN0-l3-epB" secondAttribute="centerX" id="wuX-Ph-Q16"/>
-            </constraints>
-            <nil key="simulatedStatusBarMetrics"/>
-            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
-            <connections>
-                <outlet property="tableView" destination="eZd-xX-eav" id="oQP-XR-K81"/>
-                <outlet property="tableViewHeight" destination="Ndx-EF-i13" id="FRK-eK-vaa"/>
-                <outlet property="verticalCenter" destination="dHH-SM-1al" id="R9J-vd-hNZ"/>
-            </connections>
-        </view>
-    </objects>
-</document>
