+++ intermine-ios/SlideAnimator.swift
@@ -1,0 +1,21 @@
+//
+//  SlideAnimator.swift
+//  intermine-ios
+//
+//  Created by Nadia on 5/25/17.
+//  Copyright © 2017 Nadia. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+class SlideAnimator: NSObject, UIViewControllerAnimatedTransitioning {
+    
+    func transitionDuration(using transitionContext: UIViewControllerContextTransitioning?) -> TimeInterval {
+        return 0
+    }
+    
+    func animateTransition(using transitionContext: UIViewControllerContextTransitioning) {
+        
+    }
+}
